Move `@glint/transform` into `@glint/core`
