+++ packages/core/package.json
@@ -32,7 +32,7 @@
     "typescript": "^4.7.0"
   },
   "dependencies": {
-    "@glint/transform": "^0.9.7",
+    "@glimmer/syntax": "^0.84.2",
     "escape-string-regexp": "^4.0.0",
     "minimatch": "^3.0.4",
     "silent-error": "^1.1.1",
@@ -44,11 +44,13 @@
   },
   "devDependencies": {
     "@glint/test-utils": "^0.9.7",
+    "@types/common-tags": "^1.8.0",
     "@types/minimatch": "^3.0.3",
     "@types/node": "^18.11.5",
     "@types/uuid": "^8.3.4",
     "@types/yargs": "^17.0.10",
     "@vitest/ui": "^0.23.4",
+    "common-tags": "^1.8.0",
     "execa": "^4.0.1",
     "strip-ansi": "^6.0.0",
     "vitest": "^0.22.0"
