+++ packages/core/src/transform/diagnostics/augmentation.ts
