+++ packages/core/src/transform/diagnostics/create-transform-diagnostic.ts
