+++ packages/core/src/transform/diagnostics/index.ts
