+++ packages/core/src/transform/diagnostics/rewrite-diagnostic.ts
