+++ packages/core/src/transform/template/inlining/index.ts
