+++ packages/core/src/transform/template/map-template-contents.ts
