+++ packages/core/src/transform/template/mapping-tree.ts
