+++ packages/core/src/transform/template/rewrite-module.ts
