+++ packages/core/src/transform/template/scope-stack.ts
