+++ packages/core/src/transform/template/transformed-module.ts
