+++ packages/environment-glimmerx/-private/dsl/globals.d.ts
@@ -7,12 +7,12 @@ interface Keywords {
   each: VM.EachKeyword;
   'has-block': VM.HasBlockKeyword;
   'has-block-params': VM.HasBlockParamsKeyword;
-  if: void; // the `if` keyword is implemented directly in @glint/transform
+  if: void; // the `if` keyword is implemented directly in `@glint/core`
   'in-element': VM.InElementKeyword;
   let: VM.LetKeyword;
-  unless: void; // the `unless` keyword is implemented directly in @glint/transform
+  unless: void; // the `unless` keyword is implemented directly in `@glint/core`
   with: VM.WithKeyword;
-  yield: void; // the `yield` keyword is implemented directly in @glint/transform
+  yield: void; // the `yield` keyword is implemented directly in `@glint/core`
 }
 
 export declare const Globals: Keywords & Globals;
