+++ packages/template/-private/dsl.d.ts
@@ -1,6 +1,6 @@
 // This module exports functions and types directly referenced by
-// the output of `@glint/transform`. It comprises the basis from
-// which environments construct their exports.
+// the output of `@glint/core`'s transform layer. It comprises the
+// basis from which environments construct their exports.
 
 export * from './dsl/resolve';
 export * from './dsl/emit';
