--- packages/transform/.gitignore
@@ -1,2 +1,0 @@
-lib/
-tsconfig.tsbuildinfo
