--- packages/transform/README.md
@@ -1,5 +1,0 @@
-# `@glint/transform`
-
-This package contains tools for rewriting Glimmer templates as a TypeScript approximation of their semantics, used by other [glint] packages.
-
-[glint]: https://github.com/typed-ember/glint
