--- packages/transform/__tests__/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "include": ["../src", "**/*.test.ts"],
-  "compilerOptions": {
-    "rootDir": ".."
-  }
-}
