--- packages/transform/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@glint/transform",
-  "version": "0.9.7",
-  "repository": "typed-ember/glint",
-  "description": "Transforms Glimmer templates into TypeScript code that approximates their semantics",
-  "license": "MIT",
-  "author": "Dan Freeman (https://github.com/dfreeman)",
-  "type": "module",
-  "exports": {
-    ".": "./lib/index.js"
-  },
-  "files": [
-    "README.md",
-    "lib"
-  ],
-  "scripts": {
-    "test": "vitest run",
-    "test:watch": "vitest watch",
-    "build": "tsc --build",
-    "prepack": "yarn build"
-  },
-  "dependencies": {
-    "@glimmer/syntax": "^0.84.2"
-  },
-  "devDependencies": {
-    "@types/common-tags": "^1.8.0",
-    "@types/debug": "^4.1.5",
-    "common-tags": "^1.8.0",
-    "vitest": "^0.22.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  }
-}
