+++ tsconfig.json
@@ -5,7 +5,6 @@
     { "path": "packages/environment-ember-template-imports" },
     { "path": "packages/environment-glimmerx" },
     { "path": "packages/template" },
-    { "path": "packages/transform" },
     { "path": "packages/vscode" },
     { "path": "packages/scripts" },
     { "path": "test-packages/test-utils" }
